Move PathTraversalChecker to :files
