+++ subprojects/core/src/main/java/org/gradle/internal/resource/local/DefaultPathKeyFileStore.java
@@ -24,9 +24,9 @@ import org.gradle.api.file.FileVisitDetails;
 import org.gradle.api.internal.file.collections.MinimalFileTree;
 import org.gradle.api.internal.file.collections.SingleIncludePatternFileTree;
 import org.gradle.internal.UncheckedException;
+import org.gradle.internal.file.PathTraversalChecker;
 import org.gradle.internal.hash.ChecksumService;
 import org.gradle.util.internal.GFileUtils;
-import org.gradle.util.internal.PathTraversalChecker;
 import org.gradle.util.internal.RelativePathUtil;
 
 import java.io.File;
