+++ subprojects/dependency-management/src/main/java/org/gradle/api/internal/artifacts/transform/UnzipTransform.java
@@ -38,7 +38,7 @@ import java.util.zip.ZipEntry;
 import java.util.zip.ZipInputStream;
 
 import static org.apache.commons.io.FilenameUtils.removeExtension;
-import static org.gradle.util.internal.PathTraversalChecker.safePathName;
+import static org.gradle.internal.file.PathTraversalChecker.safePathName;
 
 /**
  * Provides a generic transform from a zipped file to an extracted directory.  The extracted directory
