+++ subprojects/wrapper-shared/build.gradle.kts
@@ -10,9 +10,11 @@ gradlebuildJava.usedInWorkers()
 
 dependencies {
 
-    compileOnly(project(":base-annotations")) {
-        because("Compile only because we want to keep the wrapper.jar small")
+    implementation(project(":base-annotations"))
+    implementation(project(":files")) {
+        because("We need org.gradle.internal.file.PathTraversalChecker")
     }
+
     testImplementation(project(":base-services"))
     testImplementation(project(":core-api"))
     testImplementation(project(":native"))
